Sprint 5

See merge request I500872/babysita!4

+++ src/main/java/s3/fontys/babysita/business/impl/UserServiceImpl.java
@@ -8,11 +8,14 @@ import s3.fontys.babysita.business.UserService;
 import s3.fontys.babysita.business.exception.DuplicatedUsernameException;
 import s3.fontys.babysita.business.exception.InvalidIdException;
 import s3.fontys.babysita.business.exception.InvalidRoleException;
+import s3.fontys.babysita.business.exception.UnauthorizedDataAccessException;
 import s3.fontys.babysita.business.mapper.UserMapper;
+import s3.fontys.babysita.configuration.security.token.AccessToken;
 import s3.fontys.babysita.domain.*;
 import s3.fontys.babysita.persistence.UserRepository;
 import s3.fontys.babysita.persistence.entity.UserEntity;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -24,6 +27,7 @@ public class UserServiceImpl implements UserService {
     private final UserRepository userRepository;
     private final UserMapper userMapper;
     private final PasswordEncoder passwordEncoder;
+    private final AccessToken accessToken;
 
     @Override
     public void createUser(UserRequest userRequest, String password) {
@@ -56,6 +60,7 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteUser(int userId) {
+        checkUserPermission(userId);
         userRepository.deleteById(userId);
     }
 
@@ -80,6 +85,8 @@ public class UserServiceImpl implements UserService {
     public void partialUpdateUser(Integer id, UserRequest userUpdates) {
         UserEntity existingUser = userRepository.findById(id)
                 .orElseThrow(() -> new InvalidIdException("Invalid User id"));
+
+        checkUserPermission(id);
 
         if(userUpdates.getFirstName() != null) {
             existingUser.setFirstName(userUpdates.getFirstName());
@@ -110,6 +117,25 @@ public class UserServiceImpl implements UserService {
         return matches.stream()
                 .map(userMapper::toResponse)
                 .collect(Collectors.toList());
+    }
+
+    @Override
+    public List<UserResponse> getUsersById(int firstUserId, int secondUserId) {
+        List<UserResponse> users = new ArrayList<>();
+        UserResponse user1 = getUser(firstUserId);
+        UserResponse user2 = getUser(secondUserId);
+        users.add(user1);
+        users.add(user2);
+        return users;
+    }
+
+    @Override
+    public void checkUserPermission(int userId) {
+        if (!accessToken.getRole().equals("admin")) {
+            if (accessToken.getUserId() != userId) {
+                throw new UnauthorizedDataAccessException("USER_ID_NOT_FROM_LOGGED_IN_USER");
+            }
+        }
     }
 }
